Battle BG & Tilemap Lines temp fix
So I decided to have a go at creating the battle BG in unity using the existing tilemaps. For the moment it gives us something. Also found a potential fix for the lines.

Set the antialiasing in the project settings to disabled, this fixed about 80% of the lines. Then I decreased the grid size from 1x 1y to .9999x & y respectively which seemed to remove the problem. Garbage fix but we can find a better one later.